cleanup SQLMetrics
